requirements.txt
awsiotsdk==1.5.2
asyncio==3.4.3
